Store the player's direction in an enum instead of a pointer.
This will be necessary for saving the player's direction. We cannot
save a pointer, and we cannot use pointer equality to compute the
player's direction since we copy the Player in some places (the
p->anim pointer will point to the p->leftas or p->rightas field of the
original, not the copy). Additionally, this is more consistent with
how we handle p->act.